pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "thoughtwire"
version = "0.2.0"
description = "Binary communication protocol for AI agent networks"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{name = "flint-dominic"}]
keywords = ["ai", "agents", "mqtt", "binary", "protocol"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Communications",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "paho-mqtt>=2.0",
]

[project.optional-dependencies]
bridge = ["websockets>=12.0"]
signing = ["cryptography>=42.0"]
all = ["websockets>=12.0", "cryptography>=42.0"]

[project.scripts]
thoughtwire = "thoughtwire.cli:main"

[project.urls]
Repository = "https://github.com/flint-dominic/thoughtwire"

[tool.pytest.ini_options]
testpaths = ["tests"]
